+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - CHANGE use plain json errors inside of RxError parameters to make debugging easier.
 - CHANGE crunch multiple incremental (aka 'atomic') operations into a single database write. Also batch writes to multiple documents into a single write.
 - CHANGE Make RxDocuments immutable
+- ADD `RxDocument.getLatest()`
 
 <!-- ADD new changes here! -->